A file "registration.py"  is divided into 3 files.
