+++ apple_store/static/styles/sass/style.sass
@@ -4,7 +4,6 @@
     font-size: 24px
     color: black
 
-
 main
     height: 100%
     width: 100%
